==> source/1-factchecks/2-clean_factchecks/1-clean_factchecks.ipynb <==
import pandas as pd
from datetime import date
import datetime

def final_date_factcheck(x, y):
    if pd.isna(x):
        return(y)
    else:
        return(x)

data = pd.read_excel('/Users/cblanesg/misinformation_socialmedia/data/1-factchecks/1-posts_factchecks/factchecks_data.xlsx').drop(['Unnamed: 0', 
                                                                                                                                  'Unnamed: 15', 
                                                                                                                                  'Unnamed: 16', 
                                                                                                                                  'Unnamed: 17', 
                                                                                                                                  'Unnamed: 18'], axis = 1)

data['date_factcheck_final'] = data.apply(lambda x: final_date_factcheck(x['date_factcheck_facebook'], 
                                                                         x['fecha_chequeo']), axis = 1)

data = data[~data['date_factcheck_final'].isna()]

data['date_factcheck_final'] = data['date_factcheck_final'].apply(lambda x: x.strftime('%Y-%m-%d'))

data_factchecks = data.drop('id_desinformacion', axis = 1).drop_duplicates('id_factcheck', keep = 'first')

def partnership_function(x):
    if pd.isna(x):
        return(None)
    else:
        return(x.strftime('%Y-%m-%d'))

data_factchecks['facebook_partnership_date'] = data_factchecks['facebook_partnership_date'].apply(lambda x: partnership_function(x))

data_factchecks['fecha_chequeo'] = data_factchecks['fecha_chequeo'].apply(lambda x: partnership_function(x))

data_factchecks['date_factcheck_facebook'] = data_factchecks['date_factcheck_facebook'].apply(lambda x: partnership_function(x))

data_factchecks.to_json('../../../data/1-factchecks/2-clean_factchecks/factchecks_data.json')

### Correct Data For social Fact checks

factchecks = pd.read_json('../../../data/1-factchecks/2-clean_factchecks/factchecks_data.json')

factchecks.fecha_chequeo.unique()

def clean_dates(x):
    if int(x.split('-')[2]) > 12:
        elements = x.split('-')
        return(datetime.date(int(elements[0]), 
                            int(elements[1]), 
                            int(elements[2])))
    else:
        elements = x.split('-')
        return(datetime.date(int(elements[0]), 
                            int(elements[2]), 
                            int(elements[1])))

factchecks['fecha_factcheck'] = factchecks['date_factcheck_final'].apply(lambda x: clean_dates(x))
factchecks['fecha_factcheck'] = factchecks['fecha_factcheck'].apply(lambda x: x.strftime('%Y-%m-%d'))

factchecks.to_json('../../../data/1-factchecks/2-clean_factchecks/factchecks_data.json')

factchecks
