==> source/1-factchecks/2-clean_factchecks/2-subset_factchecks.ipynb <==
import pandas as pd

import warnings
warnings.filterwarnings("ignore")

## Functions

def load_misinformation():
    misinformations = pd.read_excel('/Users/cblanesg/misinformation_socialmedia/data/2-misinformation/newspapers/0-misinformation/misinformation_newspapers.xlsx').drop('Unnamed: 0', axis = 1)    
    return(misinformations)

def load_subset_facthecks():
    newspapers_id = pd.read_excel('../../../data/1-factchecks/2-clean_factchecks/factchecks_subset_newspapers.xlsx').dropna()
    factchecks_newspapers = list(newspapers_id.id_factcheck.unique())
    
    social_id = pd.read_excel('../../../data/1-factchecks/2-clean_factchecks/factchecks_subset_social.xlsx').dropna()
    factchecks_social = list(social_id.id_factcheck.unique())
    misisnformations = factchecks_social + factchecks_newspapers
    return(misisnformations)

def load_subset_factchecks():
    ids_factchecks = load_subset_facthecks()
    factchecks = pd.read_json('../../../data/1-factchecks/2-clean_factchecks/factchecks_data.json')
    subset_factchecks = factchecks[factchecks['id_factcheck'].isin(load_subset_facthecks())]
    return(subset_factchecks)

def save_factchecks(df):
    df.to_excel('../../../data/1-factchecks/2-clean_factchecks/subset_factchecks_data.xlsx')

## Apply Functions

subset_factchecks = load_subset_factchecks()

save_factchecks(subset_factchecks)

social_id = pd.read_excel('../../../data/1-factchecks/2-clean_factchecks/factchecks_subset_social.xlsx').dropna()
social_id = social_id[['id_desinformacion', 'id_factcheck']].drop_duplicates()

social_id[['id_desinformacion', 'id_factcheck']].drop_duplicates()
